Refactor: 포인트 관련 기능 추가

See merge request s10-webmobile1-sub2/S10P12E108!183

+++ backend/src/main/java/com/ssafy/umzip/domain/clean/dto/user/CleanCompleteReservationDto.java
@@ -10,9 +10,9 @@ import lombok.Setter;
 @NoArgsConstructor
 public class CleanCompleteReservationDto {
     private Long mappingId;
-    private Long point; //minus point
+    private int point; //minus point
     @Builder
-    public CleanCompleteReservationDto(Long mappingId, Long point) {
+    public CleanCompleteReservationDto(Long mappingId, int point) {
         this.mappingId = mappingId;
         this.point = point;
     }
